+++ packages/elvis-colors/src/index.ts
@@ -1,9 +1,4 @@
-export {
-  getThemeColor,
-  getThemeColorContrast,
-  getCustomThemeColor,
-  getBaseThemeColor,
-} from './getThemeColor';
+export { getThemeColor, getThemeColorContrast, getCustomThemeColor, getBaseColor } from './getThemeColor';
 export { ThemeName, ThemeClassName, ColorLabel } from './theme';
 
 export { shadows, getShadow } from './shadows';